3ds: improved unistore support, human readable filesizes, better zip handling
The unistore support still leaves a lot to desired, it can handle single files, but for zips it doesn't properly
respect the install scripts. To do that properly, libget's manifest engine should be refactored (a desire we have
to support local pkgbuild.json files (eg. from spinarak repo) as well).